qa_board/requirements.txt
Django>=4.2,<5.0
djangorestframework
django-cors-headers
drf-spectacular
channels[daphne]
channels-redis
celery
redis
django-celery-results
pillow
python-dotenv
